Refactor reports
