+++ metadata-ingestion/src/datahub/ingestion/source/sagemaker.py
@@ -8,44 +8,12 @@ from datahub.ingestion.source.aws_common import AwsSourceConfig
 from datahub.ingestion.source.sagemaker_processors.feature_groups import (
     FeatureGroupProcessor,
 )
+from datahub.ingestion.source.sagemaker_processors.common import (
+    SagemakerSourceConfig,
+    SagemakerSourceReport,
+)
 from datahub.ingestion.source.sagemaker_processors.jobs import JobProcessor
 from datahub.ingestion.source.sagemaker_processors.models import ModelProcessor
-
-
-class SagemakerSourceConfig(AwsSourceConfig):
-
-    extract_feature_groups: Optional[bool] = True
-    extract_models: Optional[bool] = True
-    extract_jobs: Optional[Union[Dict[str, str], bool]] = True
-
-    @property
-    def sagemaker_client(self):
-        return self.get_client("sagemaker")
-
-
-@dataclass
-class SagemakerSourceReport(SourceReport):
-    feature_groups_scanned = 0
-    features_scanned = 0
-    models_scanned = 0
-    jobs_scanned = 0
-    datasets_scanned = 0
-
-    # TODO: report these
-    def report_feature_group_scanned(self) -> None:
-        self.feature_groups_scanned += 1
-
-    def report_feature_scanned(self) -> None:
-        self.features_scanned += 1
-
-    def report_model_scanned(self) -> None:
-        self.models_scanned += 1
-
-    def report_job_scanned(self) -> None:
-        self.jobs_scanned += 1
-
-    def report_dataset_scanned(self) -> None:
-        self.datasets_scanned += 1
 
 
 class SagemakerSource(Source):
@@ -66,6 +34,7 @@ class SagemakerSource(Source):
 
     def get_workunits(self) -> Iterable[MetadataWorkUnit]:
 
+        # extract feature groups if specified
         if self.source_config.extract_feature_groups:
 
             feature_group_processor = FeatureGroupProcessor(
@@ -73,6 +42,7 @@ class SagemakerSource(Source):
             )
             yield from feature_group_processor.get_workunits()
 
+        # extract models if specified
         if self.source_config.extract_models:
 
             model_processor = ModelProcessor(
@@ -80,6 +50,7 @@ class SagemakerSource(Source):
             )
             yield from model_processor.get_workunits()
 
+        # extract jobs if specified
         if self.source_config.extract_jobs is not False:
 
             job_processor = JobProcessor(
